Make ray an optional dependency
