+++ src/pydvl/utils/parallel/__init__.py
@@ -1,3 +1,7 @@
 from .backend import *
+from .backends import *
 from .futures import *
 from .map_reduce import *
+
+if len(BaseParallelBackend.BACKENDS) == 0:
+    raise ImportError("No parallel backend found. Please install ray or joblib.")
