+++ src/pydvl/utils/parallel/backends/__init__.py
@@ -1,0 +1,9 @@
+try:
+    from .joblib import *
+except ImportError:
+    pass
+
+try:
+    from .ray import *
+except ImportError:
+    pass
